+++ src/NLightning.Infrastructure.Bitcoin/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## v0.0.4
+
+Bump version to use `NLightning.Infrastructure@v1.0.3`.
+
 ## v0.0.3
 
 Bump version to use `NLightning.Infrastructure@v1.0.2`.